Refactor pt. 2 separation of concerns
