+++ shared/hydrogen-vm-render-script.js
