+++ shared/room-directory-vm-render-script.js
